save hvg matrix as np 32 to save memory

+++ src/adata_hf_datasets/initial_embedder.py
@@ -1581,7 +1581,8 @@ class GeneSelectEmbedder(BaseEmbedder):
 
         # Get the processed expression matrix
         X = adata.X.toarray() if sp.issparse(adata.X) else adata.X
-        embedding_matrix = X.copy()
+        # Save as float32
+        embedding_matrix = X.copy().astype(np.float32)
 
         # Store in adata for compatibility
         adata.obsm[obsm_key] = embedding_matrix
